+++ consensus/debugging.go
@@ -36,12 +36,17 @@ func (m *consensusModule) SetUtilityContext(utilityContext modules.UtilityContex
 	m.utilityContext = utilityContext
 }
 
-func (m *consensusModule) SetAggregatedMetadata(maxHeight, minHeight uint64, peerAddress string) {
+func (m *consensusModule) SetAggregatedStateSyncMetadata(minHeight uint64, maxHeight uint64, peerAddress string) {
 	m.stateSync.SetAggregatedSyncMetadata(&typesCons.StateSyncMetadataResponse{
-		MaxHeight:   maxHeight,
 		MinHeight:   minHeight,
+		MaxHeight:   maxHeight,
 		PeerAddress: peerAddress,
 	})
+}
+
+func (m *consensusModule) GetAggregatedStateSyncMetadataMaxHeight() (maxHeight uint64) {
+	metadata := m.stateSync.GetAggregatedStateSyncMetadata()
+	return metadata.MaxHeight
 }
 
 func (m *consensusModule) HandleDebugMessage(debugMessage *messaging.DebugMessage) error {
@@ -156,7 +161,7 @@ func (m *consensusModule) sendGetBlockStateSyncMessage(_ *messaging.DebugMessage
 		},
 	}
 
-	validators, err := m.GetValidatorsAtHeight(currentHeight)
+	validators, err := m.getValidatorsAtHeight(currentHeight)
 	if err != nil {
 		m.logger.Debug().Msgf(typesCons.ErrPersistenceGetAllValidators.Error(), err)
 	}
@@ -186,7 +191,7 @@ func (m *consensusModule) sendGetMetadataStateSyncMessage(_ *messaging.DebugMess
 		},
 	}
 
-	validators, err := m.GetValidatorsAtHeight(currentHeight)
+	validators, err := m.getValidatorsAtHeight(currentHeight)
 	if err != nil {
 		m.logger.Debug().Msgf(typesCons.ErrPersistenceGetAllValidators.Error(), err)
 	}