+++ consensus/hotstuff_leader.go
@@ -336,7 +336,7 @@ func (m *consensusModule) validateMessageSignature(msg *typesCons.HotstuffMessag
 
 	address := partialSig.GetAddress()
 
-	validators, err := m.GetValidatorsAtHeight(m.CurrentHeight())
+	validators, err := m.getValidatorsAtHeight(m.CurrentHeight())
 	if err != nil {
 		return err
 	}