+++ shared/core/types/fsm_states.go
@@ -10,7 +10,7 @@ const (
 
 	StateMachineState_Consensus_Unsynched StateMachineState = "Consensus_Unsynched"
 	StateMachineState_Consensus_SyncMode  StateMachineState = "Consensus_SyncMode"
-	StateMachineState_Consensus_Synced    StateMachineState = "Consensus_Synced"
+	StateMachineState_Consensus_Synched   StateMachineState = "Consensus_Synched"
 
 	// Used by synched validators to participate in the HotPOKT lifecycle
 	StateMachineState_Consensus_Pacemaker StateMachineState = "Consensus_Pacemaker"