+++ shared/messaging/messages_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"testing"
 
-	//"github.com/pokt-network/pocket/consensus"
 	typesCons "github.com/pokt-network/pocket/consensus/types"
 	"github.com/pokt-network/pocket/shared/messaging"
 	typesUtil "github.com/pokt-network/pocket/utility/types"