+++ consensus/doc/CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.37] - 2023-03-07
+## [0.0.0.37] - 2023-03-08
 
 - Add `fsm_handler.go` to handle FSM transition events in consensus module
 - Update State Machine mock in `utils_test.go`
+- Update state_sync module with additional function definitions 
 
 ## [0.0.0.36] - 2023-02-28