+++ shared/CHANGELOG.md
@@ -8,10 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
-## [0.0.0.39] - 2023-03-07
+## [0.0.0.39] - 2023-03-08
 
-- Add `ConsensusFSMHandlers` interface to handle state transition events in consensus module
-- Update FSM events `Consensus_IsSynchedValidator`, `Consensus_IsSynchedNonValidator` and state `Consensus_Pacemaker`. 
+- Add `ConsensusStateSync` interface to define exported state sync functions in consensus module
+- Update `ConsensusDebugModule` with getter and setter function for state sync testing
+- Update FSM events `Consensus_IsSynchedValidator`, `Consensus_IsSynchedNonValidator` and state `Consensus_Pacemaker` 
 
 ## [0.0.0.38] - 2023-03-03